+++ src/main/java/konkuk/ptal/service/IUserService.java
@@ -1,7 +1,9 @@
 package konkuk.ptal.service;
 
 
+import konkuk.ptal.dto.request.CreateRevieweeDto;
 import konkuk.ptal.dto.request.CreateReviewerRequestDto;
+import konkuk.ptal.entity.Reviewee;
 import konkuk.ptal.entity.Reviewer;
 
 public interface IUserService {
@@ -11,4 +13,9 @@ public interface IUserService {
 
     Reviewer updateReviewer(Long id, CreateReviewerRequestDto dto, Long authenticatedUserId);
 
+    Reviewee registerReviewee(CreateRevieweeDto dto, Long authenticatedUserId);
+
+    Reviewee getReviewee(Long id);
+
+    Reviewee updateReviewee(Long id, CreateRevieweeDto dto, Long authenticatedUserId);
 }
